refactoring integration test.
