+++ integration/src/utils/instruction.rs
@@ -6,8 +6,8 @@ use crate::const_definition::{
     UDT_1_HOLDER_ACP_ADDRESS, UDT_1_HOLDER_ACP_ADDRESS_PK,
 };
 use crate::utils::address::{
-    build_acp_address, build_cheque_address, generate_rand_pw_address_pk_pair,
-    generate_rand_secp_address_pk_pair, get_udt_hash_by_owner,
+    acp::build_acp_address, cheque::build_cheque_address, generate_rand_secp_address_pk_pair,
+    get_udt_hash_by_owner,
 };
 use crate::utils::rpc_client::{CkbRpcClient, MercuryRpcClient};
 use crate::utils::signer::sign_transaction;
@@ -186,7 +186,7 @@ pub(crate) fn send_transaction_to_ckb(tx: Transaction) -> Result<H256> {
     let ckb_client = CkbRpcClient::new(CKB_URI.to_string());
     let tx_hash = ckb_client.send_transaction(tx, OutputsValidator::Passthrough)?;
     println!("send tx: 0x{}", tx_hash);
-    let _ = aggregate_transactions_into_blocks()?;
+    aggregate_transactions_into_blocks()?;
     Ok(tx_hash)
 }
 
@@ -194,41 +194,8 @@ pub(crate) fn prepare_secp_address_with_ckb_capacity(
     capacity: u64,
 ) -> Result<(Address, H256, OutPoint)> {
     let (address, pk) = generate_rand_secp_address_pk_pair();
-    let payload = TransferPayload {
-        asset_info: AssetInfo::new_ckb(),
-        from: vec![JsonItem::Address(GENESIS_BUILT_IN_ADDRESS_1.to_string())],
-        to: vec![ToInfo {
-            address: address.to_string(),
-            amount: (capacity as u128).into(),
-        }],
-        output_capacity_provider: Some(OutputCapacityProvider::From),
-        pay_fee: None,
-        fee_rate: None,
-        since: None,
-    };
-    let mercury_client = MercuryRpcClient::new(MERCURY_URI.to_string());
-    let tx = mercury_client.build_transfer_transaction(payload)?;
-    let tx = sign_transaction(tx, &[GENESIS_BUILT_IN_ADDRESS_1_PRIVATE_KEY])?;
-
-    // send tx to ckb node
-    let tx_hash = send_transaction_to_ckb(tx)?;
-
-    let tx_info = mercury_client
-        .get_transaction_info(tx_hash)?
-        .transaction
-        .expect("get transaction info");
-    let out_point = &tx_info
-        .records
-        .into_iter()
-        .find(|record| {
-            record.asset_info.asset_type == AssetType::CKB
-                && record.amount == (capacity as u128).into()
-                && record.io_type == IOType::Output
-        })
-        .expect("find record")
-        .out_point;
-
-    Ok((address, pk, out_point.to_owned()))
+    let out_point = prepare_ckb_capacity(&address, capacity)?;
+    Ok((address, pk, out_point))
 }
 
 pub(crate) fn issue_udt_with_cheque(
@@ -311,8 +278,15 @@ pub(crate) fn prepare_account(
     Ok(())
 }
 
-pub(crate) fn prepare_pw_address_with_capacity(capacity: u64) -> Result<(Address, H256, OutPoint)> {
-    let (address, pk) = generate_rand_pw_address_pk_pair();
+pub fn dump_data<T>(data: &T, file_name: &str) -> Result<()>
+where
+    T: ?Sized + Serialize,
+{
+    let json_string = serde_json::to_string_pretty(data)?;
+    std::fs::write(file_name, json_string).map_err(Into::into)
+}
+
+pub(crate) fn prepare_ckb_capacity(address: &Address, capacity: u64) -> Result<OutPoint> {
     let payload = TransferPayload {
         asset_info: AssetInfo::new_ckb(),
         from: vec![JsonItem::Address(GENESIS_BUILT_IN_ADDRESS_1.to_string())],
@@ -347,13 +321,5 @@ pub(crate) fn prepare_pw_address_with_capacity(capacity: u64) -> Result<(Address
         .expect("find record")
         .out_point;
 
-    Ok((address, pk, out_point.to_owned()))
-}
-
-pub fn dump_data<T>(data: &T, file_name: &str) -> Result<()>
-where
-    T: ?Sized + Serialize,
-{
-    let json_string = serde_json::to_string_pretty(data)?;
-    std::fs::write(file_name, json_string).map_err(Into::into)
+    Ok(out_point.to_owned())
 }
